Add store test

+++ db/sqlc/entry.sql.go
@@ -37,7 +37,7 @@ const deleteEntry = `-- name: DeleteEntry :exec
 DELETE FROM entries where id = $1
 `
 
-func (q *Queries) DeleteEntry(ctx context.Context, id int32) error {
+func (q *Queries) DeleteEntry(ctx context.Context, id int64) error {
 	_, err := q.db.ExecContext(ctx, deleteEntry, id)
 	return err
 }
@@ -47,7 +47,7 @@ SELECT id, account_id, amount, created_at FROM entries
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetEntry(ctx context.Context, id int32) (Entry, error) {
+func (q *Queries) GetEntry(ctx context.Context, id int64) (Entry, error) {
 	row := q.db.QueryRowContext(ctx, getEntry, id)
 	var i Entry
 	err := row.Scan(
@@ -106,7 +106,7 @@ WHERE id = $1
 `
 
 type UpdateEntryParams struct {
-	ID     int32 `json:"id"`
+	ID     int64 `json:"id"`
 	Amount int64 `json:"amount"`
 }
